Split out WorkflowIT.java
