+++ dockstore-integration-testing/src/test/java/io/dockstore/client/cli/WorkflowIT.java
@@ -315,89 +315,7 @@ public class WorkflowIT extends BaseIT {
             workflows.stream().allMatch(workflow -> workflow.getMode() == Workflow.ModeEnum.STUB));
     }
 
-    @Test
-    public void testTargettedRefresh() throws ApiException, URISyntaxException, IOException {
-
-        testingPostgres.runUpdateStatement("update enduser set isadmin = 't' where username = 'DockstoreTestUser2';");
-
-        final ApiClient webClient = getWebClient(USER_2_USERNAME, testingPostgres);
-        WorkflowsApi workflowApi = new WorkflowsApi(webClient);
-
-        UsersApi usersApi = new UsersApi(webClient);
-        User user = usersApi.getUser();
-        Assert.assertNotEquals("getUser() endpoint should actually return the user profile", null, user.getUserProfiles());
-
-        workflowApi.manualRegister(SourceControl.GITHUB.name(), DOCKSTORE_TEST_USER_2_HELLO_DOCKSTORE_NAME, "/Dockstore.cwl", "",
-                DescriptorLanguage.CWL.getShortName(), "/test.json");
-        workflowApi.manualRegister(SourceControl.BITBUCKET.name(), "dockstore_testuser2/dockstore-workflow", "/Dockstore.cwl", "",
-                DescriptorLanguage.CWL.getShortName(), "/test.json");
-        List<Workflow> workflows = usersApi.userWorkflows(user.getId());
-
-        for (Workflow workflow : workflows) {
-            assertNotSame("", workflow.getWorkflowName());
-        }
-
-        // do targetted refresh, should promote workflow to fully-fleshed out workflow
-        final Workflow workflowByPathGithub = workflowApi.getWorkflowByPath(DOCKSTORE_TEST_USER2_HELLO_DOCKSTORE_WORKFLOW, BIOWORKFLOW, null);
-        final Workflow refreshGithub = workflowApi.refresh(workflowByPathGithub.getId(), false);
-        final Workflow workflowByPathBitbucket = workflowApi.getWorkflowByPath(DOCKSTORE_TEST_USER2_DOCKSTORE_WORKFLOW, BIOWORKFLOW, null);
-        final Workflow refreshBitbucket = workflowApi.refresh(workflowByPathBitbucket.getId(), false);
-
-        // tests for reference type for bitbucket workflows
-        assertTrue("should see at least 4 branches", refreshBitbucket.getWorkflowVersions().stream()
-            .filter(version -> version.getReferenceType() == WorkflowVersion.ReferenceTypeEnum.BRANCH).count() >= 4);
-        assertTrue("should see at least 1 tags", refreshBitbucket.getWorkflowVersions().stream()
-            .filter(version -> version.getReferenceType() == WorkflowVersion.ReferenceTypeEnum.TAG).count() >= 1);
-
-        assertSame("github workflow is not in full mode", refreshGithub.getMode(), Workflow.ModeEnum.FULL);
-        assertTrue("github workflow version count is wrong: " + refreshGithub.getWorkflowVersions().size(),
-                4 <= refreshGithub.getWorkflowVersions().size());
-        assertEquals("should find two versions with files for github workflow, found : " + refreshGithub.getWorkflowVersions().stream()
-                .filter(workflowVersion -> !fileDAO.findSourceFilesByVersion(workflowVersion.getId()).isEmpty()).count(), 2,
-            refreshGithub.getWorkflowVersions().stream().filter(workflowVersion -> !fileDAO.findSourceFilesByVersion(workflowVersion.getId()).isEmpty()).count());
-        assertEquals("should find two valid versions for github workflow, found : " + refreshGithub.getWorkflowVersions().stream()
-                .filter(WorkflowVersion::isValid).count(), 2,
-            refreshGithub.getWorkflowVersions().stream().filter(WorkflowVersion::isValid).count());
-
-        assertSame("bitbucket workflow is not in full mode", refreshBitbucket.getMode(), Workflow.ModeEnum.FULL);
-
-        assertEquals("bitbucket workflow version count is wrong: " + refreshBitbucket.getWorkflowVersions().size(), 5,
-            refreshBitbucket.getWorkflowVersions().size());
-        assertEquals("should find 4 versions with files for bitbucket workflow, found : " + refreshBitbucket.getWorkflowVersions().stream()
-                .filter(workflowVersion -> !fileDAO.findSourceFilesByVersion(workflowVersion.getId()).isEmpty()).count(), 4,
-            refreshBitbucket.getWorkflowVersions().stream().filter(workflowVersion -> !fileDAO.findSourceFilesByVersion(workflowVersion.getId()).isEmpty()).count());
-        assertEquals("should find 0 valid versions for bitbucket workflow, found : " + refreshBitbucket.getWorkflowVersions().stream()
-                .filter(WorkflowVersion::isValid).count(), 0,
-            refreshBitbucket.getWorkflowVersions().stream().filter(WorkflowVersion::isValid).count());
-
-        // should not be able to get content normally
-        Ga4GhApi anonymousGa4Ghv2Api = new Ga4GhApi(CommonTestUtilities.getWebClient(false, null, testingPostgres));
-        Ga4GhApi adminGa4Ghv2Api = new Ga4GhApi(webClient);
-        boolean exceptionThrown = false;
-        try {
-            anonymousGa4Ghv2Api
-                .toolsIdVersionsVersionIdTypeDescriptorGet("CWL", "#workflow/" + DOCKSTORE_TEST_USER2_HELLO_DOCKSTORE_WORKFLOW, "master");
-        } catch (ApiException e) {
-            exceptionThrown = true;
-        }
-        assertTrue(exceptionThrown);
-        FileWrapper adminToolDescriptor = adminGa4Ghv2Api
-            .toolsIdVersionsVersionIdTypeDescriptorGet("CWL", "#workflow/" + DOCKSTORE_TEST_USER2_HELLO_DOCKSTORE_WORKFLOW, "testCWL");
-        assertTrue("could not get content via optional auth", adminToolDescriptor != null && !adminToolDescriptor.getContent().isEmpty());
-
-        workflowApi.publish(refreshGithub.getId(), CommonTestUtilities.createPublishRequest(true));
-        // check on URLs for workflows via ga4gh calls
-        FileWrapper toolDescriptor = adminGa4Ghv2Api
-            .toolsIdVersionsVersionIdTypeDescriptorGet("CWL", "#workflow/" + DOCKSTORE_TEST_USER2_HELLO_DOCKSTORE_WORKFLOW, "testCWL");
-        String content = IOUtils.toString(new URI(toolDescriptor.getUrl()), StandardCharsets.UTF_8);
-        assertFalse(content.isEmpty());
-        checkForRelativeFile(adminGa4Ghv2Api, "#workflow/" + DOCKSTORE_TEST_USER2_HELLO_DOCKSTORE_WORKFLOW, "testCWL", "Dockstore.cwl");
-
-        // check on commit ids for github
-        boolean allHaveCommitIds = refreshGithub.getWorkflowVersions().stream().noneMatch(version -> version.getCommitID().isEmpty());
-        assertTrue("not all workflows seem to have commit ids", allHaveCommitIds);
-    }
-
+    
     @Test
     public void testHostedEditAndDelete() {
         final ApiClient webClient = getWebClient(USER_2_USERNAME, testingPostgres);
@@ -1091,69 +1009,7 @@ public class WorkflowIT extends BaseIT {
                 WorkflowSubClass.BIOWORKFLOW.getValue()).size());
     }
 
-    /**
-     * Tests manual registration and publishing of a github and bitbucket workflow
-     *
-     * @throws ApiException exception used for errors coming back from the web service
-     */
-    @Test
-    public void testManualRegisterThenPublish() throws ApiException {
-        final ApiClient webClient = getWebClient(USER_2_USERNAME, testingPostgres);
-        WorkflowsApi workflowApi = new WorkflowsApi(webClient);
-
-        // Make publish request (true)
-        final PublishRequest publishRequest = CommonTestUtilities.createPublishRequest(true);
-
-        // Manually register workflow github
-        Workflow githubWorkflow = workflowApi
-            .manualRegister("github", DOCKSTORE_TEST_USER_2_HELLO_DOCKSTORE_NAME, "/Dockstore.wdl", "altname", "wdl", "/test.json");
-        Assert.assertEquals("test repo for CWL and WDL workflows", githubWorkflow.getTopicAutomatic());
-
-        // Manually register workflow bitbucket
-        Workflow bitbucketWorkflow = workflowApi
-            .manualRegister("bitbucket", "dockstore_testuser2/dockstore-workflow", "/Dockstore.cwl", "altname", "cwl", "/test.json");
-
-        // Assert some things
-        final long count = testingPostgres
-            .runSelectStatement("select count(*) from workflow where mode = '" + Workflow.ModeEnum.FULL + "'", long.class);
-        assertEquals("No workflows are in full mode", 0, count);
-        final long count2 = testingPostgres.runSelectStatement("select count(*) from workflow where workflowname = 'altname'", long.class);
-        assertEquals("There should be two workflows with name altname, there are " + count2, 2, count2);
-
-        // Publish github workflow
-        Workflow refreshedWorkflow = workflowApi.refresh(githubWorkflow.getId(), false);
-        workflowApi.publish(githubWorkflow.getId(), publishRequest);
-
-        // Tag with a valid descriptor (but no description) and a recognizable README
-        final String tagName = "1.0.0";
-
-        Optional<WorkflowVersion> testWDL = refreshedWorkflow.getWorkflowVersions().stream().filter(workflowVersion -> workflowVersion.getName().equals(tagName)).findFirst();
-        Assert.assertTrue("A workflow version with a descriptor that does not have a description should fall back to README", testWDL.get().getDescription().contains("test repo for CWL and WDL workflows"));
-
-        // Intentionally mess up description to test if refresh fixes it
-        testingPostgres.runUpdateStatement("update version_metadata set description='bad_potato'");
-
-        refreshedWorkflow = workflowApi.refresh(githubWorkflow.getId(), true);
-        workflowApi.publish(githubWorkflow.getId(), publishRequest);
-
-        testWDL = refreshedWorkflow.getWorkflowVersions().stream().filter(workflowVersion -> workflowVersion.getName().equals(tagName)).findFirst();
-        Assert.assertTrue("A workflow version that had a README description should get updated", testWDL.get().getDescription().contains("test repo for CWL and WDL workflows"));
-
-        // Assert some things
-        assertEquals("should have two published, found  " + workflowApi.allPublishedWorkflows(null, null, null, null, null, false, null).size(),
-            1, workflowApi.allPublishedWorkflows(null, null, null, null, null, false, null).size());
-        final long count3 = testingPostgres
-            .runSelectStatement("select count(*) from workflow where mode = '" + Workflow.ModeEnum.FULL + "'", long.class);
-        assertEquals("One workflow is in full mode", 1, count3);
-        final long count4 = testingPostgres.runSelectStatement("select count(*) from workflowversion where valid = 't'", long.class);
-        assertTrue("There should be at least 2 valid version tags, there are " + count4, 2 <= count4);
-
-        workflowApi.refresh(bitbucketWorkflow.getId(), false);
-        thrown.expect(ApiException.class);
-        // Publish bitbucket workflow, will fail now since the workflow test case is actually invalid now
-        workflowApi.publish(bitbucketWorkflow.getId(), publishRequest);
-    }
-
+   
     /**
      * Tests that the info for quay images included in CWL workflows are grabbed and that the trs endpoints convert this info correctly
      */
